second

+++ ThreadPool/TimerThread.cpp
@@ -1,6 +1,6 @@
 #include "TimerThread.h"
 
-TimerThread::TimerThread(int val, int interval, Callback cb)
+TimerThread::TimerThread(double val, double interval, Callback cb)
 	:_timer(val, interval, std::move(cb)),
 	 _thread(std::bind(&Timer::start, &_timer))
 {
